POCs/3RFID/requirements.txt
RPi.GPIO
pirc522
requests